altered max-window button
This is code designed to be added to a custom post type in Wordpress (one that is totally blank except for the chart, allowing easy iframe access to said chart). 
I updated the contextButton 'symbol' to reference an image in the theme/images folder. To do this I had to update the child theme's functions.php file as per this article: http://codex.wordpress.org/Function_Reference/wp_localize_script
I simply co-opted one of the highchart script files I was enqueuing for adding a localized variable 'blankchart_template_url' and as it is an array (sheepishly copying the example exactly), I put the template url in 'blankchart_template_url.template_url'. This is now called within the javascript here. 
Sadly, this template url points to the parent theme - so I've added the 'images/maximize-window-20.png to the parent theme rather than my child theme.